Refactor pages and categories
