+++ app/watch/WatchPageClient.tsx
@@ -1,136 +1,186 @@
 "use client"
+
 import Link from "next/link"
+import { useRouter, useSearchParams } from "next/navigation"
+import { useState, useEffect } from "react"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { Play, Eye, Calendar, Youtube, Clock, Video } from "lucide-react"
-import { useState, useEffect } from "react"
+import { Play, Eye, Calendar, Youtube, Clock, Video, Star, ArrowRight } from "lucide-react"
+import { CategoryTabs } from "@/components/ui/category-tabs"
+import dynamic from "next/dynamic"
 
-interface VideoType {
+const CategoryDropdown = dynamic(() => import("@/components/ui/category-dropdown").then(mod => mod.CategoryDropdown), { ssr: false })
+
+interface Video {
   id: string
   title: string
-  slug: string
-  description: string
-  youtubeId: string
-  thumbnail: string
-  duration: string
-  category: string
-  tags: string[]
+  category?: string
   status: string
-  publishedAt: string
-  views: number
-  featured: boolean
+  featured?: boolean
+  views?: number
+  duration?: string
 }
 
-interface Category {
-  id: string
+interface CategoryData {
   name: string
   slug: string
-  color?: string
+  count: number
 }
 
-export function WatchPageClient({
-  searchParams,
-}: {
-  searchParams: { category?: string }
-}) {
-  const [videos, setVideos] = useState<VideoType[]>([])
-  const [categories, setCategories] = useState<Category[]>([])
-  const [loading, setLoading] = useState(true)
-  const selectedCategory = searchParams.category || "all"
+interface WatchPageProps {
+  initialVideos: Video[]
+  initialCategories: CategoryData[]
+}
 
-  useEffect(() => {
-    fetchWatchData()
-  }, [])
+export function WatchPageClient({ initialVideos, initialCategories }: WatchPageProps) {
+  const router = useRouter()
+  const searchParams = useSearchParams()
+  const categoryParam = searchParams?.get("category") || "all"
+  const pageParam = parseInt(searchParams?.get("page") || "1", 10)
 
-  const fetchWatchData = async () => {
-    try {
-      setLoading(true)
-      const [videosResponse, categoriesResponse] = await Promise.all([
-        fetch("/api/admin/videos", { cache: "no-store" }),
-        fetch("/api/admin/categories?type=video", { cache: "no-store" }),
-      ])
+  const [selectedCategory, setSelectedCategory] = useState(categoryParam)
+  const [currentPage, setCurrentPage] = useState(pageParam)
+  const [videos] = useState(initialVideos)
+  const [categories] = useState(initialCategories)
 
-      if (videosResponse.ok) {
-        const videosData = await videosResponse.json()
-        const publishedVideos = videosData.videos?.filter((video: VideoType) => video.status === "published") || []
-        setVideos(publishedVideos)
-      }
+  const videosPerPage = 12
 
-      if (categoriesResponse.ok) {
-        const categoriesData = await categoriesResponse.json()
-        setCategories(categoriesData.categories || [])
-      }
-    } catch (error) {
-      console.error("Failed to fetch watch data:", error)
-    } finally {
-      setLoading(false)
-    }
+  // Normalize category slug for filtering consistency
+  const normalizedSelectedCategory = selectedCategory
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/(^-|-$)/g, "")
+
+  // Filter videos by category
+  const filteredVideos =
+    normalizedSelectedCategory === "all"
+      ? videos
+      : videos.filter((video) => {
+          const categorySlug = video.category
+            ?.toLowerCase()
+            .replace(/[^a-z0-9]+/g, "-")
+            .replace(/(^-|-$)/g, "")
+          return categorySlug === normalizedSelectedCategory
+        })
+
+  const totalPages = Math.ceil(filteredVideos.length / videosPerPage)
+  const startIndex = (currentPage - 1) * videosPerPage
+  const paginatedVideos = filteredVideos.slice(startIndex, startIndex + videosPerPage)
+
+  // Separate featured & regular videos
+  const featuredVideos = filteredVideos.filter((video) => video.featured).slice(0, 3)
+  const regularVideos = paginatedVideos.filter((video) => !video.featured)
+
+  // Stats
+  const totalViews = videos.reduce((sum, video) => sum + (video.views || 0), 0)
+  const totalDuration = videos.reduce((sum, video) => {
+    const duration = video.duration || "0:00"
+    const [minutes, seconds] = duration.split(":").map(Number)
+    return sum + (minutes || 0) + (seconds || 0) / 60
+  }, 0)
+  const avgDuration = videos.length > 0 ? Math.round(totalDuration / videos.length) : 0
+
+  // Client-side category change
+  function handleCategoryChange(newCategory: string) {
+    setSelectedCategory(newCategory)
+    setCurrentPage(1)
+    router.push(`/watch?category=${newCategory}&page=1`, { scroll: true })
   }
 
-  const filteredVideos =
-    selectedCategory === "all" ? videos : videos.filter((video) => video.category === selectedCategory)
+  // Client-side page change
+  function handlePageChange(newPage: number) {
+    setCurrentPage(newPage)
+    router.push(`/watch?category=${selectedCategory}&page=${newPage}`, { scroll: true })
+  }
 
-  const featuredVideos = filteredVideos.filter((video) => video.featured)
-  const regularVideos = filteredVideos.filter((video) => !video.featured)
+  const EmptyState = ({ selectedCategory }: { selectedCategory: string }) => (
+    <div className="text-center py-12 sm:py-16 lg:py-20">
+      <div className="max-w-md mx-auto px-4">
+        <div className="w-20 h-20 sm:w-24 sm:h-24 bg-gradient-to-br from-blue-100 to-purple-100 rounded-full flex items-center justify-center mx-auto mb-6">
+          <Video className="h-10 w-10 sm:h-12 sm:w-12 text-blue-600" />
+        </div>
+        <h3 className="text-xl sm:text-2xl font-bold text-gray-900 mb-4">
+          {selectedCategory === "all" ? "No videos yet" : `No videos in this category`}
+        </h3>
+        <p className="text-gray-600 mb-6 sm:mb-8 leading-relaxed text-sm sm:text-base">
+          {selectedCategory === "all"
+            ? "We're working on exciting video content about anime creation, tutorials, and behind-the-scenes footage. Check back soon!"
+            : `No videos have been published in this category yet. Try selecting a different category or check back later.`}
+        </p>
+        <div className="bg-gradient-to-br from-blue-50 to-purple-50 border border-blue-200 rounded-lg p-4 sm:p-6">
+          <p className="text-blue-800 text-sm">
+            🎬 <strong>Content creators:</strong> Use the admin panel to upload your first video!
+          </p>
+        </div>
+      </div>
+    </div>
+  )
 
   return (
     <div className="min-h-screen bg-gray-50">
-      {/* Beautiful Header - Match screenshot style */}
+      {/* Header */}
       <div className="bg-white border-b">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-20">
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-12 sm:py-16 lg:py-20">
           <div className="text-center">
-            {/* Blue Badge */}
-            <div className="inline-flex items-center gap-2 bg-blue-600 text-white rounded-full px-6 py-2 mb-6 text-sm font-medium">
+            <div className="inline-flex items-center gap-2 bg-blue-600 text-white rounded-full px-4 sm:px-6 py-2 mb-4 sm:mb-6 text-sm font-medium shadow-lg">
               <Video className="h-4 w-4" />
               Original Content & Series
             </div>
-
-            {/* Gradient Heading */}
-            <h1 className="text-5xl md:text-6xl font-bold mb-6 text-transparent bg-clip-text bg-gradient-to-r from-blue-600 via-purple-600 to-blue-800">
+            <h1 className="text-3xl sm:text-4xl md:text-5xl lg:text-6xl font-bold mb-4 sm:mb-6 text-transparent bg-clip-text bg-gradient-to-r from-blue-600 via-purple-600 to-blue-800 leading-relaxed px-4 pb-2">
               Watch Our Stories
             </h1>
-
-            {/* Italic Subtitle */}
-            <p className="text-xl md:text-2xl text-gray-600 italic font-light max-w-3xl mx-auto">
+            <p className="text-base sm:text-lg md:text-xl lg:text-2xl text-gray-600 italic font-light max-w-3xl mx-auto mb-6 sm:mb-8 px-4">
               Bringing anime to life—one frame at a time.
             </p>
-
-            {/* Stats */}
-            <div className="mt-8 flex items-center justify-center gap-6 text-sm text-gray-500">
-              <span className="flex items-center gap-2">
-                <span className="w-2 h-2 bg-red-500 rounded-full"></span>
-                {videos.length} Videos Available
-              </span>
-              <span>•</span>
-              <span className="flex items-center gap-2">
-                <Youtube className="h-4 w-4" />
-                Subscribe for Updates
-              </span>
+            <div className="grid grid-cols-2 lg:grid-cols-4 gap-3 sm:gap-4 max-w-2xl mx-auto px-4">
+              <div className="bg-gradient-to-br from-blue-50 to-purple-50 rounded-lg p-3 sm:p-4 border border-blue-100">
+                <div className="text-xl sm:text-2xl font-bold text-blue-600">{videos.length}</div>
+                <div className="text-xs sm:text-sm text-gray-600">Videos</div>
+              </div>
+              <div className="bg-gradient-to-br from-purple-50 to-pink-50 rounded-lg p-3 sm:p-4 border border-purple-100">
+                <div className="text-xl sm:text-2xl font-bold text-purple-600">{categories.length}</div>
+                <div className="text-xs sm:text-sm text-gray-600">Categories</div>
+              </div>
+              <div className="bg-gradient-to-br from-green-50 to-blue-50 rounded-lg p-3 sm:p-4 border border-green-100">
+                <div className="text-xl sm:text-2xl font-bold text-green-600">{totalViews.toLocaleString()}</div>
+                <div className="text-xs sm:text-sm text-gray-600">Total Views</div>
+              </div>
+              <div className="bg-gradient-to-br from-orange-50 to-red-50 rounded-lg p-3 sm:p-4 border border-orange-100">
+                <div className="text-xl sm:text-2xl font-bold text-orange-600">{avgDuration}</div>
+                <div className="text-xs sm:text-sm text-gray-600">Avg Duration</div>
+              </div>
             </div>
           </div>
         </div>
       </div>
 
-      {/* Category Filter */}
-      <section className="bg-gray-50 border-b">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-6">
-          <div className="flex flex-col sm:flex-row items-center justify-between gap-4">
-            <h3 className="text-lg font-semibold text-gray-900">Browse by Category</h3>
-            <CategoryFilter categories={categories} selectedCategory={selectedCategory} />
-          </div>
-        </div>
-      </section>
+      {/* Category Dropdown */}
+      <CategoryDropdown
+        categories={categories}
+        selectedCategory={selectedCategory}
+        onCategoryChange={handleCategoryChange}
+      />
+
+      {/* Category Tabs */}
+      <CategoryTabs
+        categories={[{ name: "All", slug: "all" }, ...categories]}
+        selectedCategory={selectedCategory}
+        basePath="/watch"
+        type="video"
+      />
 
       {/* Featured Videos */}
       {featuredVideos.length > 0 && (
-        <section className="py-16">
+        <section className="py-8 sm:py-12 lg:py-16 bg-gradient-to-br from-gray-50 to-blue-50">
           <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <h2 className="text-3xl font-bold text-gray-900 mb-8">Featured Content</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-              {featuredVideos.slice(0, 2).map((video) => (
-                <FeaturedVideoCard key={video.id} video={video} />
+            <div className="flex items-center gap-2 mb-6 sm:mb-8">
+              <Star className="h-5 w-5 sm:h-6 sm:w-6 text-yellow-500" />
+              <h2 className="text-xl sm:text-2xl lg:text-3xl font-bold text-gray-900">Featured Videos</h2>
+            </div>
+            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 sm:gap-6 lg:gap-8">
+              {featuredVideos.map((video, index) => (
+                <FeaturedVideoCard key={video.id} video={video} priority={index === 0} />
               ))}
             </div>
           </div>
@@ -138,51 +188,63 @@ export function WatchPageClient({
       )}
 
       {/* All Videos */}
-      <section className="py-16 bg-white">
+      <section className="py-8 sm:py-12 lg:py-16 bg-white">
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex items-center justify-between mb-8">
-            <h2 className="text-3xl font-bold text-gray-900">
-              {selectedCategory === "all" ? "Latest Videos" : `${selectedCategory} Videos`}
-            </h2>
-            <div className="flex items-center gap-4">
-              <span className="text-sm text-gray-500">
-                {filteredVideos.length} video{filteredVideos.length !== 1 ? "s" : ""}
-              </span>
-              <Button asChild variant="outline">
-                <Link href="https://www.youtube.com/@flavorstudios" target="_blank">
-                  <Youtube className="mr-2 h-4 w-4" />
-                  View All on YouTube
-                </Link>
-              </Button>
+          <div className="flex flex-col sm:flex-row items-start sm:items-center justify-between mb-6 sm:mb-8 gap-4">
+            <div>
+              <h2 className="text-xl sm:text-2xl lg:text-3xl font-bold text-gray-900 mb-2">
+                {selectedCategory === "all"
+                  ? "Latest Videos"
+                  : `${categories.find((c) => c.slug === selectedCategory)?.name || selectedCategory} Videos`}
+              </h2>
+              <p className="text-gray-600 text-sm sm:text-base">
+                {filteredVideos.length} video{filteredVideos.length !== 1 ? "s" : ""} found
+              </p>
             </div>
+            {totalPages > 1 && (
+              <div className="text-sm text-gray-500">
+                Page {currentPage} of {totalPages}
+              </div>
+            )}
           </div>
 
-          {loading ? (
-            <VideosSkeleton />
-          ) : filteredVideos.length === 0 ? (
+          {filteredVideos.length === 0 ? (
             <EmptyState selectedCategory={selectedCategory} />
           ) : (
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-6">
-              {regularVideos.map((video) => (
-                <VideoCard key={video.id} video={video} />
-              ))}
-            </div>
+            <>
+              <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4 sm:gap-6 lg:gap-8 mb-8 sm:mb-12">
+                {regularVideos.map((video) => (
+                  <VideoCard key={video.id} video={video} />
+                ))}
+              </div>
+
+              {/* Pagination */}
+              {totalPages > 1 && (
+                <Pagination
+                  currentPage={currentPage}
+                  totalPages={totalPages}
+                  selectedCategory={selectedCategory}
+                  onPageChange={handlePageChange}
+                />
+              )}
+            </>
           )}
         </div>
       </section>
 
-      {/* YouTube Channel CTA */}
-      <section className="py-20 bg-gradient-to-r from-blue-600 to-purple-600 text-white">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
-          <h2 className="text-3xl font-bold mb-4">Subscribe to Our YouTube Channel</h2>
-          <p className="text-xl mb-8 opacity-90">
+      {/* YouTube CTA */}
+      <section className="py-12 sm:py-16 bg-gradient-to-r from-blue-600 to-purple-600 text-white">
+        <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
+          <h2 className="text-xl sm:text-2xl lg:text-3xl font-bold mb-4">Subscribe to Our YouTube Channel</h2>
+          <p className="text-base sm:text-lg lg:text-xl mb-6 sm:mb-8 opacity-90">
             Don't miss any of our latest content! Subscribe for weekly episodes, behind-the-scenes content, and anime
             news.
           </p>
-          <Button asChild size="lg" variant="secondary">
-            <Link href="https://www.youtube.com/@flavorstudios" target="_blank">
-              <Youtube className="mr-2 h-5 w-5" />
+          <Button asChild size="lg" variant="secondary" className="shadow-lg">
+            <Link href="https://www.youtube.com/@flavorstudios" target="_blank" rel="noopener noreferrer">
+              <Youtube className="mr-2 h-4 w-4 sm:h-5 sm:w-5" />
               Subscribe on YouTube
+              <ArrowRight className="ml-2 h-4 w-4 sm:h-5 sm:w-5" />
             </Link>
           </Button>
         </div>
@@ -191,176 +253,3 @@ export function WatchPageClient({
   )
 }
 
-function FeaturedVideoCard({ video }: { video: VideoType }) {
-  const thumbnailUrl = video.thumbnail || `https://img.youtube.com/vi/${video.youtubeId}/maxresdefault.jpg`
-
-  return (
-    <Link href={`/watch/${video.slug || video.id}`}>
-      <Card className="overflow-hidden hover:shadow-xl transition-all duration-300 hover:-translate-y-1">
-        <div className="relative">
-          <img src={thumbnailUrl || "/placeholder.svg"} alt={video.title} className="w-full h-64 object-cover" />
-          <div className="absolute inset-0 bg-black bg-opacity-40 flex items-center justify-center opacity-0 hover:opacity-100 transition-opacity">
-            <Button size="lg" className="bg-blue-600 hover:bg-blue-700">
-              <Play className="mr-2 h-5 w-5" />
-              Watch Now
-            </Button>
-          </div>
-          <div className="absolute bottom-4 right-4 bg-black bg-opacity-75 text-white px-2 py-1 rounded text-sm flex items-center gap-1">
-            <Clock className="h-3 w-3" />
-            {video.duration}
-          </div>
-          <Badge className="absolute top-4 left-4 bg-yellow-500">Featured</Badge>
-        </div>
-        <CardHeader>
-          <Badge variant="secondary" className="w-fit mb-2">
-            {video.category}
-          </Badge>
-          <CardTitle className="text-xl line-clamp-2">{video.title}</CardTitle>
-          <p className="text-gray-600 line-clamp-2">{video.description}</p>
-        </CardHeader>
-        <CardContent>
-          <div className="flex items-center justify-between text-sm text-gray-500">
-            <div className="flex items-center gap-4">
-              <span className="flex items-center gap-1">
-                <Eye className="h-4 w-4" />
-                {video.views.toLocaleString()} views
-              </span>
-              <span className="flex items-center gap-1">
-                <Calendar className="h-4 w-4" />
-                {new Date(video.publishedAt).toLocaleDateString()}
-              </span>
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-    </Link>
-  )
-}
-
-function VideoCard({ video }: { video: VideoType }) {
-  const thumbnailUrl = video.thumbnail || `https://img.youtube.com/vi/${video.youtubeId}/maxresdefault.jpg`
-
-  return (
-    <Link href={`/watch/${video.slug || video.id}`} className="group">
-      <Card className="overflow-hidden transition-all duration-300 hover:shadow-xl hover:-translate-y-2 group-hover:shadow-blue-500/25">
-        <div className="relative">
-          <img
-            src={thumbnailUrl || "/placeholder.svg"}
-            alt={video.title}
-            className="w-full h-48 object-cover transition-transform duration-300 group-hover:scale-110"
-          />
-          <div className="absolute bottom-2 right-2 bg-black/80 text-white px-2 py-1 rounded-md text-sm flex items-center gap-1 backdrop-blur-sm">
-            <Clock className="h-3 w-3" />
-            {video.duration}
-          </div>
-          <div className="absolute inset-0 flex items-center justify-center opacity-0 group-hover:opacity-100 transition-all duration-300 bg-black/50 backdrop-blur-sm">
-            <div className="bg-blue-600 rounded-full p-4 transform scale-75 group-hover:scale-100 transition-transform duration-300 shadow-2xl">
-              <Play className="h-8 w-8 text-white ml-1" />
-            </div>
-          </div>
-        </div>
-        <CardHeader className="pb-2">
-          <Badge variant="secondary" className="w-fit mb-2 border-blue-200 text-blue-700 bg-blue-50">
-            {video.category}
-          </Badge>
-          <CardTitle className="line-clamp-2 text-lg group-hover:text-blue-600 transition-colors">
-            {video.title}
-          </CardTitle>
-        </CardHeader>
-        <CardContent>
-          <div className="flex items-center justify-between text-sm text-gray-500">
-            <div className="flex items-center gap-1">
-              <Calendar className="h-4 w-4" />
-              {new Date(video.publishedAt).toLocaleDateString()}
-            </div>
-            <div className="flex items-center gap-1">
-              <Eye className="h-4 w-4" />
-              {video.views.toLocaleString()}
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-    </Link>
-  )
-}
-
-function VideosSkeleton() {
-  return (
-    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-6">
-      {[...Array(8)].map((_, i) => (
-        <Card key={i}>
-          <div className="h-48 bg-gray-200 animate-pulse"></div>
-          <CardHeader>
-            <div className="h-4 bg-gray-200 animate-pulse rounded mb-2"></div>
-            <div className="h-6 bg-gray-200 animate-pulse rounded"></div>
-          </CardHeader>
-          <CardContent>
-            <div className="h-4 bg-gray-200 animate-pulse rounded"></div>
-          </CardContent>
-        </Card>
-      ))}
-    </div>
-  )
-}
-
-function EmptyState({ selectedCategory }: { selectedCategory: string }) {
-  return (
-    <div className="text-center py-20">
-      <div className="max-w-md mx-auto">
-        <h3 className="text-2xl font-bold text-gray-900 mb-4">
-          {selectedCategory === "all" ? "No videos yet" : `No videos in ${selectedCategory}`}
-        </h3>
-        <p className="text-gray-600 mb-6">
-          {selectedCategory === "all"
-            ? "Come back soon for new updates! We're creating amazing original anime content, tutorials, and behind-the-scenes videos."
-            : `No videos have been published in the ${selectedCategory} category yet. Try selecting a different category or check back later.`}
-        </p>
-        <div className="bg-blue-50 border border-blue-200 rounded-lg p-4 mb-6">
-          <p className="text-blue-800 text-sm">
-            🎬 <strong>Content creators:</strong> Use the admin panel to add your first video!
-          </p>
-        </div>
-        <Button asChild>
-          <Link href="https://www.youtube.com/@flavorstudios" target="_blank">
-            <Youtube className="mr-2 h-4 w-4" />
-            Subscribe on YouTube
-          </Link>
-        </Button>
-      </div>
-    </div>
-  )
-}
-
-interface CategoryFilterProps {
-  categories: Category[]
-  selectedCategory: string
-}
-
-function CategoryFilter({ categories, selectedCategory }: CategoryFilterProps) {
-  return (
-    <Select
-      value={selectedCategory}
-      onValueChange={(value) => {
-        const url = new URL(window.location.href)
-        if (value === "all") {
-          url.searchParams.delete("category")
-        } else {
-          url.searchParams.set("category", value)
-        }
-        window.location.href = url.toString()
-      }}
-    >
-      <SelectTrigger className="w-full sm:w-[200px]">
-        <SelectValue placeholder="All Categories" />
-      </SelectTrigger>
-      <SelectContent>
-        <SelectItem value="all">All Categories</SelectItem>
-        {categories.map((category) => (
-          <SelectItem key={category.id} value={category.name}>
-            {category.name}
-          </SelectItem>
-        ))}
-      </SelectContent>
-    </Select>
-  )
-}
